requirements.txt
pandas>=1.5.0
numpy>=1.20.0
requests>=2.27.1
python-dotenv>=0.20.0
matplotlib>=3.5.0
seaborn>=0.11.2
scipy>=1.8.0
streamlit>=1.20.0
pillow>=9.0.0
schwab-py>=1.0.0
ib_async>=1.0.0 